# internal/api/wellknown/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "See docs/10-running-in-production.md for token issuance. A route with no matching scope in the token, and no admin role, is refused with 403.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Every route declared in internal/apispec, the package cmd/controller/main.go also builds its real router from, served under /api/v1. A listed route's method, pattern, scope, and link relation are the ones the server enforces. That a listed route is registered at all is not checked by anything, so confirm a route against a running controller before building on it.",
    "title": "Pleiades control plane API",
    "version": "unreleased"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/teams": {
      "get": {
        "description": "Returns teams, keyset-paged. Each is a group of users, and the only thing a role is ever granted to.",
        "operationId": "list_teams",
        "parameters": [
          {
            "description": "Keyset cursor: the last id of the previous page.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Maximum records to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive name filter.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "teams": {
                      "items": {
                        "properties": {
                          "_links": {
                            "description": "Hypermedia affordances available on this resource for the calling identity.",
                            "items": {
                              "properties": {
                                "href": {
                                  "type": "string"
                                },
                                "method": {
                                  "type": "string"
                                },
                                "rel": {
                                  "type": "string"
                                }
                              },
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "attested_at": {
                            "description": "When that confirmation was made. Absent means never, which an access review must treat as stale.",
                            "format": "date-time",
                            "type": "string"
                          },
                          "attested_by": {
                            "description": "Who last confirmed this team's ownership information is current. Set only by the attest endpoint, never by a write.",
                            "type": "string"
                          },
                          "description": {
                            "description": "What this team is responsible for, as opposed to who is currently in it.",
                            "type": "string"
                          },
                          "id": {
                            "description": "Stable numeric identifier.",
                            "type": "integer"
                          },
                          "name": {
                            "description": "Unique within its organization.",
                            "type": "string"
                          },
                          "organization": {
                            "description": "The tenancy boundary this team belongs to.",
                            "type": "integer"
                          },
                          "organization_name": {
                            "description": "That tenant's name, so a client can render the reference without a second request.",
                            "type": "string"
                          },
                          "users": {
                            "description": "Member user ids.",
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "A page of teams."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No identity on the request context."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The caller lacks access:read."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:read"
            ]
          }
        ],
        "summary": "List teams",
        "x-link-relation": "collection"
      },
      "post": {
        "description": "Creates a group of users, and the only thing a role is ever granted to.",
        "operationId": "create_team",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "attested_at": {
                      "description": "When that confirmation was made. Absent means never, which an access review must treat as stale.",
                      "format": "date-time",
                      "type": "string"
                    },
                    "attested_by": {
                      "description": "Who last confirmed this team's ownership information is current. Set only by the attest endpoint, never by a write.",
                      "type": "string"
                    },
                    "description": {
                      "description": "What this team is responsible for, as opposed to who is currently in it.",
                      "type": "string"
                    },
                    "id": {
                      "description": "Stable numeric identifier.",
                      "type": "integer"
                    },
                    "name": {
                      "description": "Unique within its organization.",
                      "type": "string"
                    },
                    "organization": {
                      "description": "The tenancy boundary this team belongs to.",
                      "type": "integer"
                    },
                    "organization_name": {
                      "description": "That tenant's name, so a client can render the reference without a second request.",
                      "type": "string"
                    },
                    "users": {
                      "description": "Member user ids.",
                      "items": {
                        "type": "integer"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The created record."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The submission is not a valid team."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "That name is already taken."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:write"
            ]
          }
        ],
        "summary": "Create a team",
        "x-link-relation": "create"
      }
    },
...
    "/users": {
      "get": {
        "description": "Returns users, keyset-paged. Each is an identity, joined to a token's subject by its email address.",
        "operationId": "list_users",
        "parameters": [
          {
            "description": "Keyset cursor: the last id of the previous page.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Maximum records to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive name filter.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "users": {
                      "items": {
                        "properties": {
                          "_links": {
                            "description": "Hypermedia affordances available on this resource for the calling identity.",
                            "items": {
                              "properties": {
                                "href": {
                                  "type": "string"
                                },
                                "method": {
                                  "type": "string"
                                },
                                "rel": {
                                  "type": "string"
                                }
                              },
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "email": {
                            "description": "The join key against a token's subject. Lowercased on write.",
                            "type": "string"
                          },
                          "id": {
                            "description": "Stable numeric identifier.",
                            "type": "integer"
                          },
                          "teams": {
                            "description": "Team ids this identity belongs to.",
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "A page of users."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No identity on the request context."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The caller lacks access:read."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:read"
            ]
          }
        ],
        "summary": "List users",
        "x-link-relation": "collection"
      },
      "post": {
        "description": "Creates an identity, joined to a token's subject by its email address.",
        "operationId": "create_user",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "email": {
                      "description": "The join key against a token's subject. Lowercased on write.",
                      "type": "string"
                    },
                    "id": {
                      "description": "Stable numeric identifier.",
                      "type": "integer"
                    },
                    "teams": {
                      "description": "Team ids this identity belongs to.",
                      "items": {
                        "type": "integer"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The created record."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The submission is not a valid user."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "That name is already taken."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:write"
            ]
          }
        ],
        "summary": "Create a user",
        "x-link-relation": "create"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the controller's own origin.",
      "url": "/api/v1"
    }
  ]
}